Improve Dictionary Writer Performance
Dictionary Writers are significantly slower than the Direct writer,
even when taking cost of Dictionary maintenance into account.
This change tries to close the gap. It has variety of improvements.

1. DictionaryBuilder is only used for non null elements.
Dictionary Writer represents the null by an invalid index(-1).
2. DictionaryBuilder does not expose the Block. It is used
as a pure dictionary. This change opens up the possibility of
using non-sequential buffers to store the dictionary.
3. On abandoning String dictionary, a dictionary block creation
is avoided. Creating dictionary blocks and creating slice
per row, produces too much garbage. Slices with offset and lengths
are directly written to the SliceDirectColumnWriter.
4. Statistics will not be recomputed when abandoning String
dictionary. Existing statistics computed by String dictionary writer
will be reused.
5. RowGroupIndexes are converted from IntBigArray to int[]. IntBigArray
uses 1024 segments (4 KB) and default stride size is 10_000. So only
8 of them could be filled at a time and wasting 4KB per stride.
This on some tests account for 20% of the memory usage.
6. Rehashing of dictionary, when not empty can assume collision.
Previously rehashing was verifying the element equals.